Cleanup – Remove configs package (#701)
- Delete tiferet/configs/ (settings.py, logging.py, tests/ and all contents)
- Remove dead 'from ..configs import *' wildcard from contexts/cache.py
- Inline LegacyTiferetError class into commands/core.py (add json import)
- Remove "tiferet.configs" from pyproject.toml packages include list
