--- tiferet/configs/__init__.py
@@ -1,5 +1,0 @@
-# *** imports
-
-# ** app
-from .settings import *
-from ..models.settings import *
