--- tiferet/configs/tests/test_calc_cli.yml
@@ -1,59 +1,0 @@
-cli:
-  cmds:
-    test_calc.add_number: 
-      group_key: test-calc
-      key: add-number
-      description: Adds two numbers.
-      args:
-        - name_or_flags: 
-            - a
-          description: The first number to add.
-        - name_or_flags: 
-            - b
-          description: The second number to add.
-      name: Add Number Command
-    test_calc.subtract_number:
-      group_key: test-calc
-      key: subtract-number
-      description: Subtracts two numbers.
-      args:
-        - name_or_flags: 
-            - a
-          description: The number to subtract from.
-        - name_or_flags: 
-            - b
-          description: The number to subtract.
-      name: Subtract Number Command
-    test_calc.multiply_number:
-      group_key: test-calc
-      key: multiply-number
-      description: Multiplies two numbers.
-      args:
-        - name_or_flags: 
-            - a
-          description: The first number to multiply.
-        - name_or_flags: 
-            - b
-          description: The second number to multiply.  
-      name: Multiply Number Command
-    test_calc.divide_number:
-      group_key: test-calc
-      key: divide-number
-      description: Divides two numbers.
-      args:
-        - name_or_flags: 
-            - a
-          description: The numerator.
-        - name_or_flags: 
-            - b
-          description: The denominator.
-      name: Divide Number Command
-    test_calc.square_number:
-      group_key: test-calc
-      key: square-number
-      description: Squares a number.
-      args:
-        - name_or_flags: 
-            - a
-          description: The number to square.
-      name: Square Number Command
